+++ lib/best_in_place/railtie.rb
@@ -1,5 +1,7 @@
+require 'jquery/rails/railtie'
+
 module BestInPlace
-  class Railtie < Rails::Railtie
+  class Railtie < ::Rails::Railtie
     config.after_initialize do
       BestInPlace::ViewHelpers = ActionView::Base.new
     end