+++ lib/best_in_place/utils.rb
@@ -1,22 +1,22 @@
 module BestInPlace
   module Utils
     module_function
+    extend ActionController::ModelNaming
 
     def build_best_in_place_id(object, field)
       case object
-      when Symbol, String
-        "best_in_place_#{object}_#{field}"
-      else
-        id = "best_in_place_#{object_to_key(object)}"
-        id << "_#{object.id}" if object.class.ancestors.include?(ActiveModel::Serializers::JSON)
-        id << "_#{field}"
-        id
+        when Symbol, String
+          "best_in_place_#{object}_#{field}"
+        else
+          id = "best_in_place_#{object_to_key(object)}"
+          id << "_#{object.id}" if object.persisted?
+          id << "_#{field}"
+          id
       end
     end
 
     def object_to_key(object)
-      return object.class.to_s.underscore unless object.class.respond_to?(:model_name)
-      ActiveModel::Naming.param_key(object.class)
+      model_name_from_record_or_class(object).param_key
     end
   end
 end