+++ spec/internal/config/initializers/development.rb
@@ -1,0 +1,5 @@
+Rails.application.configure do
+  config.cache_classes = false
+  config.eager_load = false
+  config.assets.debug = true
+end