+++ spec/internal/config/routes.rb
@@ -1,7 +1,6 @@
 Rails.application.routes.draw do
   resources :users do
     member do
-      put :test_respond_with
       get :double_init
       get :show_ajax
       get :email_field